Move service factory into /llm/services folder
