+++ src/utils/service-initializer.ts
@@ -1,7 +1,7 @@
 import { ClientManager } from '../client/manager.js';
 import { ILLMService } from '../ai/llm/types.js';
 import { AgentConfig } from '../config/types.js';
-import { createLLMService } from '../ai/llm/factory.js';
+import { createLLMService } from '../ai/llm/services/factory.js';
 import { logger } from './logger.js';
 
 /**
